requirements.txt
Flask==3.0.0
numpy==1.26.1
Pillow==10.1.0
opencv-python==4.8.1.78
mtcnn==0.1.1
tensorflow==2.14.0
keras-facenet==0.3.2
scipy==1.11.3
